added filter in new mark themed modules builder plugin https://online.sbis.ru/opendoc.html?guid=5d7552c9-b770-457a-9ccb-4da76ee15245 Also removed old deprecated info about themes from readme https://online.sbis.ru/opendoc.html?guid=cc87b127-b75d-42de-9b4b-0b24cae0a62d
