+++ README.md
@@ -194,13 +194,7 @@ _Значение по умолчанию:_ **true**.\
 ![GitHub Logo](/images/symlinks-false.png)
 Флаг установлен в значение _true_:
 ![GitHub Logo](/images/symlinks.png)
-#### 17) themes
-Сборка темизируемых css.\
-_Принимаемые значения:_ **false/true**.\
-_Значение по умолчанию:_ **false**.\
-Для набора исходников(на примере Controls/Button) результат работы сборщика будет иметь вид:
-![GitHub Logo](/images/themes.png)
-#### 18) tsc
+#### 17) tsc
 Выполнение в сборке проекта команды компилятора typescript - tsc с флагом --noEmit
 (компиляция typescript без сохранения результатов компиляции - для выявления ошибок в typescript-исходниках вашего проекта)\
 _Принимаемые значения:_ **false/true**.\
